Allow `apns-topic` to be set on the client
We didn't previously have a way to set the topic, and were instead
relying on APNS to default to the topic from the certificate subject.

However, in some cases we will need to override that.  For some reason,
our newly generated certificates are not working without the topic being
explicitly specified (APNS doesn't seem to be taking it from the
subject, although it was doing it for our older certificates).

This commit adds a way to set the topic on a per-client basis, using a
new `topic` keyword argument.