Ensure socket not in invalid state after timeout
Depending on when a connection timeout occurred, it was possible to be
left with a socket that reported itself `connected?`, even though it
wasn't valid.

Changed to discard the socket completely in the case that a timeout
occurs, to be sure we start over with a fresh connection next time.